+++ CompulsoryCow.IsImplemented/CompulsoryCow.IsImplemented/Verify.cs
@@ -4,9 +4,8 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Reflection;
 
-namespace CompulsoryCow.IsEqualsImplemented;
+namespace CompulsoryCow.IsImplemented;
 
-[Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
 public class Verify
 {
     private readonly IList<Type> _ignoredClasses = new List<Type>();
@@ -36,21 +35,18 @@ public class Verify
     /// <summary>This property contains the Type of the class that does not have Equals implemented correctly.
     /// Updated every time <see cref="AreAllEqualsImplementedCorrectly(Assembly){T}"/> is called.
     /// </summary>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public Type? ResultClass { get; private set; }
 
     /// <summary>This property contains a friendly message about which property was not used in the Equals comparison. 
     /// Updated every time <see cref="IsEqualsImplementedCorrectly{T}"/> or <see cref="AreAllEqualsImplementedCorrectly(Assembly)"/> is called.
     /// If last check returned true this property is null.
     /// </summary>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public string ResultMessage { get; private set; } = "";
 
     /// <summary>This property contains the <see cref="PropertyInfo"/> for the property that was not in the Equals comparison.
     /// Updated every time <see cref="IsEqualsImplementedCorrectly{T}"/> is called.
     /// If last check returned true this property is null.
     /// </summary>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public PropertyInfo? ResultProperty { get; private set; }
 
     /// <summary>This method is used for verifying classes 
@@ -61,7 +57,6 @@ public class Verify
     /// <example>
     /// verify.AddInstantiator(() => new Customer("myName"))
     /// </example>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public void AddInstantiator<T>(Func<object> instantiateObjectAction)
     {
         _instantiateObjectActions.Add(
@@ -72,7 +67,6 @@ public class Verify
     /// <summary>This method is used for exclude classes from comparison test.
     /// </summary>
     /// <typeparam name="T"></typeparam>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public void AddIgnoredClass<T>()
     {
         _ignoredClasses.Add(typeof(T));
@@ -86,7 +80,6 @@ public class Verify
     /// </summary>
     /// <param name="assembly"></param>
     /// <returns></returns>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public bool AreAllEqualsImplementedCorrectly([DisallowNull] Assembly assembly)
     {
         if (assembly == null) throw new ArgumentNullException(nameof(assembly));
@@ -155,7 +148,6 @@ public class Verify
     /// </summary>
     /// <typeparam name="T"></typeparam>
     /// <returns></returns>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public bool HasEqualsBeenDeclared<T>()
     {
         return HasEqualsBeenDeclared(typeof(T));
@@ -172,7 +164,7 @@ public class Verify
     /// a method for explicitly instantiating the class has to be provided
     /// through <see cref="AddInstantiator{T}(Func{object})"/>.
     /// 
-    /// If the class has a property of a type that is not handled by <see cref="IsEqualsImplemented"/>, 
+    /// If the class has a property of a type that is not handled by <see cref="IsImplemented"/>, 
     /// just call <see cref="SetComparisonValues{T}(T, T)"/> for every such type.
     /// 
     /// Note: It is important to implemented GetHashCode and override == and != but this is not checked for.
@@ -183,7 +175,6 @@ public class Verify
     /// </summary>
     /// <typeparam name="T"></typeparam>
     /// <returns></returns>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public bool IsEqualsImplementedCorrectly<T>()
     {
         return IsEqualsImplementedCorrectly(typeof(T));
@@ -194,7 +185,6 @@ public class Verify
     /// <typeparam name="T"></typeparam>
     /// <param name="equalValue"></param>
     /// <param name="differingValue"></param>
-    [Obsolete("Deprecated. Use CompulsoryCow.IsImplemented instead.", false)]
     public void SetComparisonValues<T>([DisallowNull] T equalValue, [DisallowNull] T differingValue)
     {
         _equalValues.Add(typeof(T), equalValue);