+++ CompulsoryCow.IsImplemented/Tests/IsEqualsImplementedAssemblyNoDefinitionAtAll/AClassWithEqualsNotDefined.cs
@@ -1,0 +1,6 @@
+﻿namespace IsEqualsImplementedAssemblyNotOk;
+
+public class AClassWithEqualsNotDefined
+{
+    public int MyProperty { get; set; }
+}