+++ CompulsoryCow.IsImplemented/Tests/IsEqualsImplementedAssemblyOk/AClassWithEqualsNotDefined.cs
@@ -1,0 +1,6 @@
+﻿namespace IsEqualsImplementedAssemblyOk;
+
+public class AClassWithEqualsNotDefined
+{
+    public int MyProperty { get; set; }
+}