+++ CompulsoryCow.IsImplemented/CompulsoryCow.IsImplemented/CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.0]
-Breaking change: Deprecate this package. Move to IsImplemented.
+## [0.7.0]
+Breaking change: Renamed project to IsImplemented.
 Refactoring to get rid of compiler warnings.
+
+## [0.6.0]
+This version number is reserved for the last, deprecated CompulsoryCow.IsEqualsImplemented.
 
 ## [0.5.0]
 `Constructor.IsDefaultImplemented` returns whether a class contains a default constructor.